requirements.txt
# AI舌诊智能诊断系统 - 依赖包列表
# ML Engineer Agent
# 2026-02-11

# ============================================================
# 深度学习框架
# ============================================================
paddlepaddle-gpu==2.6.0  # GPU版本，如果只有CPU则使用 paddlepaddle==2.6.0
# paddlepaddle==2.6.0    # CPU版本

# ============================================================
# 数据处理
# ============================================================
pycocotools>=2.0.6       # COCO格式数据处理
albumentations>=1.3.0    # 数据增强
opencv-python>=4.8.0     # 图像处理
Pillow>=10.0.0          # 图像IO

# ============================================================
# 科学计算
# ============================================================
numpy>=1.24.0           # 数值计算
scipy>=1.11.0           # 科学计算
scikit-learn>=1.3.0     # 机器学习工具
pandas>=2.0.0           # 数据分析

# ============================================================
# 可视化
# ============================================================
matplotlib>=3.7.0        # 绘图
seaborn>=0.12.0         # 统计可视化

# ============================================================
# 实验跟踪
# ============================================================
mlflow>=2.8.0           # 实验跟踪
# wandb>=0.15.0         # 可选：WandB集成

# ============================================================
# 配置和工具
# ============================================================
pyyaml>=6.0             # YAML配置文件
tqdm>=4.65.0            # 进度条
python-dateutil>=2.8.0  # 日期处理

# ============================================================
# 模型部署（可选）
# ============================================================
# onnx>=1.14.0          # ONNX导出
# onnxruntime>=1.15.0   # ONNX推理
# paddle2onnx>=1.0.0    # Paddle转ONNX

# ============================================================
# 开发工具（可选）
# ============================================================
# pytest>=7.4.0         # 测试框架
# black>=23.0.0         # 代码格式化
# flake8>=6.0.0         # 代码检查
# ipython>=8.14.0       # 交互式Python
# jupyter>=1.0.0        # Jupyter Notebook
